+++ src/main/java/br/com/uniamerica/Pizzaria/Entity/Funcionario.java
@@ -10,18 +10,18 @@ public class Funcionario {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "id_funcionario")
-    private int id_funcionario;
+    private Long id_funcionario;
 
     @Getter @Setter
     @Column(name = "nome")
     private String nome;
 
     @Column(name = "id_usuario")
-    private int id_usuario;
+    private Long id_usuario;
 
     public Funcionario(){}
 
-    public Funcionario(int id_funcionario, String nome, int id_usuario) {
+    public Funcionario(Long id_funcionario, String nome, Long id_usuario) {
         this.id_funcionario = id_funcionario;
         this.nome = nome;
         this.id_usuario = id_usuario;
